changing location of edge node and value types to protocol module. ALL TESTS STILL PASS
