+++ helix-engine/src/graph_core/graph_core.rs
@@ -1,7 +1,7 @@
 use std::str;
 
 use crate::graph_core::graph_methods::GraphMethods;
-use crate::types::{Edge, GraphError, Node, Value};
+use crate::types::GraphError;
 use crate::HelixGraphStorage;
 pub struct HelixGraphEngine {
     storage: HelixGraphStorage,
