+++ protocol/Cargo.toml
@@ -4,12 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-
+serde = { version = "1.0", features = ["derive"] }
 
 [lib]
 name = "protocol"
 path = "src/lib.rs"
-
-[[bin]]
-name = "protocol"
-path = "src/main.rs"
